Version Packages (#13)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @lightdotso/background@0.0.1

### Patch Changes

-   90f3724: Init version release

## @lightdotso/content@0.0.1

### Patch Changes

-   90f3724: Init version release

## @lightdotso/ethereum@0.0.1

### Patch Changes

-   90f3724: Init version release

## @lightdotso/inpage@0.0.1

### Patch Changes

-   90f3724: Init version release

## @lightdotso/popup@0.0.1

### Patch Changes

-   90f3724: Init version release

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Shun Kakinoki <shunkakinoki@gmail.com>